+++ lib/src/backends_registry.dart
@@ -1,7 +1,7 @@
 import 'package:activatory/src/activation_context.dart';
 import 'package:activatory/src/aliases/type_alias_registry.dart';
 import 'package:activatory/src/backend_store.dart';
-import 'package:activatory/src/backend_store_key.dart';
+import 'package:activatory/src/resolve_key.dart';
 import 'package:activatory/src/backends/array_backend.dart';
 import 'package:activatory/src/backends/generator_backend.dart';
 import 'package:activatory/src/backends/map_backend.dart';
@@ -35,7 +35,7 @@ class BackendsRegistry {
       key = key.runtimeType;
     }
     var affectedType = _aliasesRegistry.getAlias(type);
-    var storeKey = new BackendStoreKey(affectedType, key);
+    var storeKey = new ResolveKey(affectedType, key);
     var backends = _store.find(storeKey);
     if (backends == null) {
       backends = _factory.create(affectedType);
@@ -43,7 +43,7 @@ class BackendsRegistry {
       return get(affectedType, context);
     }
 
-    var customization = _customizationsRegistry.get(affectedType);
+    var customization = _customizationsRegistry.get(affectedType, key: key);
     var ctorResolveStrategy = _ctorResolveStrategyFactory.get(customization.resolutionStrategy);
     var backend = ctorResolveStrategy.resolve(backends, context);
     return backend;
@@ -51,7 +51,7 @@ class BackendsRegistry {
 
   GeneratorBackend register(GeneratorBackend backend, Type type, {Object key}) {
     var wrapped = new RecursionLimiter(type, backend);
-    _store.store(wrapped, new BackendStoreKey(type, key));
+    _store.store(wrapped, new ResolveKey(type, key));
     return wrapped;
   }